pi/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vibesensor"
version = "0.1.0"
description = "Offline Raspberry Pi vibration telemetry server"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.111,<1",
  "uvicorn[standard]>=0.30,<1",
  "numpy>=1.26,<3",
  "PyYAML>=6,<7",
]

[project.optional-dependencies]
dev = [
  "pytest>=8,<9",
  "pytest-asyncio>=0.23,<1",
  "ruff>=0.6",
  "websockets>=12,<16",
]

[project.scripts]
vibesensor-server = "vibesensor.app:main"

[tool.setuptools]
packages = ["vibesensor"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

